refactor: deepen Email2FAService and collapse email-2fa-policy.ts
